+++ users/users-router.js
@@ -1,18 +1,11 @@
 const express = require('express');
 const userDb = require('../data/helpers/userDb.js');
 
+// Custom Middleware.
+const checkForName = require('../middlewares/checkForName.js');
+
 //  Router created
 const router = express.Router();
-
-// Custom middleware
-function checkForName(req, res, next) {
-  const {name} = req.body;
-  if (!name) {
-    res.status(400).json({errorMessage: 'Please provide a name'});
-  } else {
-    next();
-  }
-} // Checks if name was provided in the request body of a POST or PUT request
 
 router.post('/', checkForName, (req, res) => {
   const newUser = req.body;
